chore(documentation): Added sub-navigation for forms module
