+++ packages/forms/examples/forms/pages/datepicker/datepicker.page.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+	templateUrl: './datepicker.page.html',
+})
+export class FormsDatepickerDemoPageComponent {
+}
