+++ packages/forms/examples/forms/pages/demo/demo.page.html
@@ -1,206 +1,17 @@
 <h1 class="h3 u-margin-bottom">Forms</h1>
 
-<!-- Autocomplete -->
-<h2 class="h4 u-margin-bottom">Autocomplete</h2>
+<nav class="u-margin-bottom-lg">
+	<ul class="m-nav-tabs m-nav-tabs__layout">
+		<li><a [routerLink]="['autocomplete']" [routerLinkActive]="['is-active']">Autocomplete</a></li>
+		<li><a [routerLink]="['datepicker']" [routerLinkActive]="['is-active']">Datepicker</a></li>
+		<li><a [routerLink]="['field-errors']" [routerLinkActive]="['is-active']">Field errors</a></li>
+		<li><a [routerLink]="['mask']" [routerLinkActive]="['is-active']">Mask</a></li>
+		<li><a [routerLink]="['range-slider']" [routerLinkActive]="['is-active']">Range slider</a></li>
+		<li><a [routerLink]="['search-filter']" [routerLinkActive]="['is-active']">Search filter</a></li>
+		<li><a [routerLink]="['timepicker']" [routerLinkActive]="['is-active']">Timepicker</a></li>
+		<li><a [routerLink]="['upload']" [routerLinkActive]="['is-active']">Upload</a></li>
+		<li><a [routerLink]="['wysiwyg']" [routerLinkActive]="['is-active']">WYSIWYG</a></li>
+	</ul>
+</nav>
 
-<div class="u-margin-bottom-lg">
-	<aui-code-snippet [codeSnippet]="autocompleteImportExample"></aui-code-snippet>
-</div>
-
-<h3 class="h5 u-margin-bottom">Local search</h3>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="autocompleteExampleJS1"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="autocompleteExampleHTML1"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-lg">
-	<div class="a-input">
-		<div class="a-input__wrapper">
-			<aui-auto-complete
-				id="hero-names"
-				placeholder="Choose your hero…"
-				label="name"
-				value=""
-				minCharacters = "3"
-				clearInvalid="true"
-				showAllByDefault="true"
-				[data]="heroList"
-				(select)="setSelectedHero($event)">
-			</aui-auto-complete>
-		</div>
-	</div>
-</div>
-
-<h3 class="h5 u-margin-bottom">Remote search</h3>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="autocompleteExampleJS2"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="autocompleteExampleHTML2"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-xx">
-	<div class="a-input">
-		<div class="a-input__wrapper">
-			<aui-auto-complete
-				id="hero"
-				placeholder="Choose your hero…"
-				remote="true"
-				[results]="results"
-				label="name"
-				key="id"
-				loadingText = "Loading…"
-				noResultsText="No results found"
-				searchIncentiveText="Type one or more keywords to start searching"
-				(search)="searchItems($event)">
-			</aui-auto-complete>
-		</div>
-	</div>
-</div>
-
-
-<!-- Mask -->
-<h2 class="h4 u-margin-bottom">Mask</h2>
-
-<div class="u-margin-bottom">
-    <aui-code-snippet [codeSnippet]="maskImportExample"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-lg">
-    <aui-code-snippet [codeSnippet]="maskExampleHTML1"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-xx">
-    <div class="a-input">
-		<input
-			type="text"
-			placeholder="BE99 9999 9999 9999"
-			auiMask="BE99 9999 9999 9999" />
-	</div>
-</div>
-
-
-<!-- Range slider -->
-<h2 class="h4 u-margin-bottom">Range slider</h2>
-
-<div class="u-margin-bottom-lg">
-	<aui-code-snippet [codeSnippet]="rangesliderImportExample"></aui-code-snippet>
-</div>
-
-<h3 class="h5 u-margin-bottom">Basic</h3>
-
-<div class="u-margin-bottom-lg">
-	<aui-code-snippet [codeSnippet]="rangesliderExampleHTML1"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-lg">
-	<aui-range-slider></aui-range-slider>
-</div>
-
-<h3 class="h5 u-margin-bottom">Step</h3>
-
-<div class="u-margin-bottom-lg">
-	<aui-code-snippet [codeSnippet]="rangesliderExampleHTML2"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-lg">
-	<aui-range-slider
-		step="20"
-		labelAfter="%">
-	</aui-range-slider>
-</div>
-
-<h3 class="h5 u-margin-bottom">Min/max</h3>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="rangesliderExampleTypescript3"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-lg">
-	<aui-code-snippet [codeSnippet]="rangesliderExampleHTML3"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom-xx col-md-6 col-xs-12">
-	<aui-range-slider
-		[(ngModel)]="slider3"
-		min="300"
-		max="600"
-		labelBefore="€">
-	</aui-range-slider>
-</div>
-
-
-<!-- Timepicker -->
-<h2 class="h4 u-margin-bottom">Timepicker</h2>
-
-<div class="u-margin-bottom-lg">
-	<aui-code-snippet [codeSnippet]="timepickerImportExample"></aui-code-snippet>
-</div>
-
-<h3 class="h5 u-margin-bottom">Basic</h3>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="timepickerExampleTypescript"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="timepickerExampleHTML1"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-timepicker
-		size="small"
-		[(ngModel)]="time1">
-	</aui-timepicker>
-</div>
-
-<div class="u-margin-bottom-lg">
-	<pre class="a-pre">{{ time1 }}</pre>
-</div>
-
-<h3 class="h5 u-margin-bottom">Predefined hours and minutes</h3>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="timepickerExampleTypescript2"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="timepickerExampleHTML2"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-timepicker
-		hoursPlaceholder="20"
-		minutesPlaceholder="30"
-		size="small"
-		[(ngModel)]="time2">
-	</aui-timepicker>
-</div>
-
-<div class="u-margin-bottom-lg">
-	<pre class="a-pre">{{ time2 }}</pre>
-</div>
-
-<h3 class="h5 u-margin-bottom">Validation</h3>
-
-<div class="u-margin-bottom">
-	<aui-code-snippet [codeSnippet]="timepickerExampleHTML3"></aui-code-snippet>
-</div>
-
-<div class="u-margin-bottom">
-	<aui-timepicker
-		hasError="true"
-		size="large"
-		[(ngModel)]="time3">
-	</aui-timepicker>
-</div>
-
-<div class="u-margin-bottom">
-	<pre class="a-pre">{{ time3 }}</pre>
-</div>
+<router-outlet></router-outlet>
