+++ packages/forms/examples/forms/pages/field-errors/field-errors.page.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+	templateUrl: './field-errors.page.html',
+})
+export class FormsFieldErrorsDemoPageComponent {
+}
