+++ packages/forms/examples/forms/pages/index.ts
@@ -1,5 +1,23 @@
 import { FormsDemoPageComponent } from './demo/demo.page';
+import { FormsAutocompleteDemoPageComponent } from './autocomplete/autocomplete.page';
+import { FormsDatepickerDemoPageComponent } from './datepicker/datepicker.page';
+import { FormsFieldErrorsDemoPageComponent } from './field-errors/field-errors.page';
+import { FormsMaskDemoPageComponent } from './mask/mask.page';
+import { FormsRangeSliderDemoPageComponent } from './range-slider/range-slider.page';
+import { FormsSearchFilterDemoPageComponent } from './search-filter/search-filter.page';
+import { FormsTimepickerDemoPageComponent } from './timepicker/timepicker.page';
+import { FormsUploadDemoPageComponent } from './upload/upload.page';
+import { FormsWysiwygDemoPageComponent } from './wysiwyg/wysiwyg.page';
 
 export const Pages = [
 	FormsDemoPageComponent,
+	FormsAutocompleteDemoPageComponent,
+	FormsDatepickerDemoPageComponent,
+	FormsFieldErrorsDemoPageComponent,
+	FormsMaskDemoPageComponent,
+	FormsRangeSliderDemoPageComponent,
+	FormsSearchFilterDemoPageComponent,
+	FormsTimepickerDemoPageComponent,
+	FormsUploadDemoPageComponent,
+	FormsWysiwygDemoPageComponent,
 ];
