+++ packages/forms/examples/forms/pages/mask/mask.page.html
@@ -1,0 +1,18 @@
+<h2 class="h4 u-margin-bottom">Mask</h2>
+
+<div class="u-margin-bottom">
+    <aui-code-snippet [codeSnippet]="maskImportExample"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom-lg">
+    <aui-code-snippet [codeSnippet]="maskExampleHTML1"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom-xx">
+    <div class="a-input">
+		<input
+			type="text"
+			placeholder="BE99 9999 9999 9999"
+			auiMask="BE99 9999 9999 9999" />
+	</div>
+</div>
