+++ packages/forms/examples/forms/pages/range-slider/range-slider.page.html
@@ -1,0 +1,47 @@
+<h2 class="h4 u-margin-bottom">Range slider</h2>
+
+<div class="u-margin-bottom-lg">
+	<aui-code-snippet [codeSnippet]="rangesliderImportExample"></aui-code-snippet>
+</div>
+
+<h3 class="h5 u-margin-bottom">Basic</h3>
+
+<div class="u-margin-bottom-lg">
+	<aui-code-snippet [codeSnippet]="rangesliderExampleHTML1"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom-lg">
+	<aui-range-slider></aui-range-slider>
+</div>
+
+<h3 class="h5 u-margin-bottom">Step</h3>
+
+<div class="u-margin-bottom-lg">
+	<aui-code-snippet [codeSnippet]="rangesliderExampleHTML2"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom-lg">
+	<aui-range-slider
+		step="20"
+		labelAfter="%">
+	</aui-range-slider>
+</div>
+
+<h3 class="h5 u-margin-bottom">Min/max</h3>
+
+<div class="u-margin-bottom">
+	<aui-code-snippet [codeSnippet]="rangesliderExampleTypescript3"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom-lg">
+	<aui-code-snippet [codeSnippet]="rangesliderExampleHTML3"></aui-code-snippet>
+</div>
+
+<div class="u-margin-bottom-xx col-md-6 col-xs-12">
+	<aui-range-slider
+		[(ngModel)]="mySlider"
+		min="300"
+		max="600"
+		labelBefore="€">
+	</aui-range-slider>
+</div>
