+++ packages/forms/examples/forms/pages/search-filter/search-filter.page.html
@@ -1,0 +1,1 @@
+<h2 class="h4 u-margin-bottom">Search filter</h2>
