+++ packages/forms/examples/forms/pages/search-filter/search-filter.page.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+	templateUrl: './search-filter.page.html',
+})
+export class FormsSearchFilterDemoPageComponent {
+}
