+++ packages/forms/examples/forms/pages/upload/upload.page.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+	templateUrl: './upload.page.html',
+})
+export class FormsUploadDemoPageComponent {
+}
