+++ packages/forms/examples/forms/pages/wysiwyg/wysiwyg.page.html
@@ -1,0 +1,1 @@
+<h2 class="h4 u-margin-bottom">WYSIWYG</h2>
