+++ packages/forms/examples/forms/pages/wysiwyg/wysiwyg.page.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+	templateUrl: './wysiwyg.page.html',
+})
+export class FormsWysiwygDemoPageComponent {
+}
