+++ packages/layout/examples/layout/layout.module.ts
@@ -1,5 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
 import {
 	CookieconsentModule,
 	FooterModule,
@@ -11,7 +12,6 @@ import {
 } from '@acpaas-ui/ngx-components/layout';
 import { LogoModule } from '@acpaas-ui/ngx-components/logo';
 import { CodeSnippetModule } from '@acpaas-ui/ngx-components/code-snippet';
-import { RouterModule } from '@angular/router';
 
 import { AUIDemoModalComponent } from './pages/modal/demo-modal.component';
